BCK-10 Postavljanje inicijalnih paketa i klasa
