+++ user-service/src/main/java/rs/edu/raf/userservice/bootstrap/BootstrapData.java
@@ -1,0 +1,4 @@
+package rs.edu.raf.userservice.bootstrap;
+
+public class BootstrapData {
+}
