+++ user-service/src/main/java/rs/edu/raf/userservice/configuration/SpringSecurityConfig.java
@@ -1,0 +1,4 @@
+package rs.edu.raf.userservice.configuration;
+
+public class SpringSecurityConfig {
+}
