+++ user-service/src/main/java/rs/edu/raf/userservice/controllers/UserController.java
@@ -1,0 +1,4 @@
+package rs.edu.raf.userservice.controllers;
+
+public class UserController {
+}
