+++ user-service/src/main/java/rs/edu/raf/userservice/domains/exceptions/ForbiddenException.java
@@ -1,0 +1,4 @@
+package rs.edu.raf.userservice.domains.exceptions;
+
+public class ForbiddenException {
+}
