+++ user-service/src/main/java/rs/edu/raf/userservice/domains/mappers/UserMapper.java
@@ -1,0 +1,4 @@
+package rs.edu.raf.userservice.domains.mappers;
+
+public class UserMapper {
+}
