+++ user-service/src/main/java/rs/edu/raf/userservice/domains/model/Permission.java
@@ -1,4 +1,4 @@
-package rs.edu.raf.userservice.model;
+package rs.edu.raf.userservice.domains.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
@@ -6,35 +6,33 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
-import rs.edu.raf.userservice.model.enums.RoleName;
+import rs.edu.raf.userservice.domains.model.enums.PermissionName;
 
-import java.io.Serializable;
 import java.util.ArrayList;
-
 import java.util.List;
-
 
 @Data
 @Builder
 @AllArgsConstructor
 @RequiredArgsConstructor
 @Entity
-@Table(name="roles")
-public class Role implements Serializable {
+@Table(name="permissions")
+public class Permission {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long roleId;
+    private Long permissionId;
 
     @Enumerated(EnumType.STRING)
-    private RoleName roleName;
+    private PermissionName permissionName;
 
     @JsonIgnore
     @ManyToMany
     @JoinTable(
-            name = "user_roles",
-            joinColumns = @JoinColumn(name = "roleId"),
+            name = "user_permissions",
+            joinColumns = @JoinColumn(name = "permissionId"),
             inverseJoinColumns = @JoinColumn(name = "userId")
     )
     private List<User> users = new ArrayList<>();
+
 }
