+++ user-service/src/main/java/rs/edu/raf/userservice/domains/model/User.java
@@ -1,4 +1,4 @@
-package rs.edu.raf.userservice.model;
+package rs.edu.raf.userservice.domains.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
@@ -9,9 +9,7 @@ import lombok.*;
 
 import java.io.Serializable;
 import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 
 @Data
 @Builder
