+++ user-service/src/main/java/rs/edu/raf/userservice/filters/JwtFilter.java
@@ -1,0 +1,4 @@
+package rs.edu.raf.userservice.filters;
+
+public class JwtFilter {
+}
