+++ user-service/src/main/java/rs/edu/raf/userservice/repositories/RoleRepository.java
@@ -1,0 +1,4 @@
+package rs.edu.raf.userservice.repositories;
+
+public interface RoleRepository {
+}
