+++ user-service/src/main/java/rs/edu/raf/userservice/services/RoleService.java
@@ -1,0 +1,4 @@
+package rs.edu.raf.userservice.services;
+
+public class RoleService {
+}
