+++ user-service/src/main/java/rs/edu/raf/userservice/services/UserService.java
@@ -1,0 +1,4 @@
+package rs.edu.raf.userservice.services;
+
+public class UserService {
+}
