+++ user-service/src/main/java/rs/edu/raf/userservice/utils/JwtUtil.java
@@ -1,0 +1,4 @@
+package rs.edu.raf.userservice.utils;
+
+public class JwtUtil {
+}
